Flattened the SBM and Network classes together.
